Added all the Math and Turtle Commands. Some require some more scrutiny such as the clearscreen command because I am unsure what property holds for making the lines. In addition, I changed the interface to be return arg a number instead of specifically an integer.

+++ src/slogo/Model/Commands/MathOperations/ProductCommand.java
@@ -32,8 +32,8 @@ public class ProductCommand extends Command {
   }
 
   @Override
-  public Integer returnArgValue() {
-    return (int) this.returnArgValue;
+  public Double returnArgValue() {
+    return this.returnArgValue;
   }
 
 }
